requirements.txt
python-dotenv
google-adk
google-generativeai
